add article editor
